// src/types.ts
export interface CertificateData {
  livro: string;
  folha: string;
  nubente1Nome: string;
  nubente2Nome: string;
  dataCasamento: string;
  municipioCasamento: string;
  estadoCasamento: string;
  pastorNome: string;
  pastorCpf: string;
  testemunha1Nome: string;
  testemunha1Endereco: string;
  testemunha1Cidade: string;
  testemunha1Estado: string;
  testemunha2Nome: string;
  testemunha2Endereco: string;
  testemunha2Cidade: string;
  testemunha2Estado: string;
  nubente1Nacionalidade: string;
  nubente1NaturalidadeMunicipio: string;
  nubente1NaturalidadeEstado: string;
  nubente1DataNascimento: string;
  nubente1EstadoCivil: string;
  nubente1Profissao: string;
  nubente1Domicilio: string;
  nubente1EstadoDomicilio: string;
  nubente1Pai: string;
  nubente1Mae: string;
  nubente2Nacionalidade: string;
  nubente2NaturalidadeMunicipio: string;
  nubente2NaturalidadeEstado: string;
  nubente2DataNascimento: string;
  nubente2EstadoCivil: string;
  nubente2Profissao: string;
  nubente2Domicilio: string;
  nubente2EstadoDomicilio: string;
  nubente2Pai: string;
  nubente2Mae: string;
  emissaoCidade: string;
  emissaoEstadoUf: string;
  emissaoData: string;
}